+++ language/move-command-line-common/Cargo.toml
@@ -12,10 +12,12 @@ edition = "2021"
 [dependencies]
 anyhow = "1.0.52"
 difference = "2.0.0"
+dirs-next = "2.0.0"
 walkdir = "2.3.1"
 sha2 = "0.9.3"
 hex = "0.4.3"
 num-bigint = "0.4.0"
+once_cell = "1.7.2"
 serde = { version = "1.0.124", features = ["derive"] }
 
 move-core-types = { path = "../move-core/types" }
