+++ language/move-command-line-common/src/movey.rs
@@ -1,0 +1,8 @@
+// Copyright (c) The Diem Core Contributors
+// Copyright (c) The Move Contributors
+// SPDX-License-Identifier: Apache-2.0
+
+#[cfg(debug_assertions)]
+pub const MOVEY_URL: &str = "https://movey-app-staging.herokuapp.com";
+#[cfg(not(debug_assertions))]
+pub const MOVEY_URL: &str = "https://www.movey.net";
